created RequestController to hold delete,put,post, and get functions. I can correctly call all 4 methods. I still need to configure Put, Delete, and get
